Create End-Point Enzymatic Assay Analysis
This script performs data cleaning and analysis for the End-Point Enzymatic Assay according to the determined peak areas corresponding to product ans substrate, including filtering, renaming, and calculating ratios of substrate to product areas. It also generates a bar plot of the ratios and saves it as a PNG file.